% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{download_single_zip_file}
\alias{download_single_zip_file}
\title{Downloads a single zip file from SEC}
\usage{
download_single_zip_file(year_in, quarter_in,
  cache_folder = "GetEdgarData-cache")
}
\arguments{
\item{year_in}{The year}

\item{quarter_in}{The quarter (e.g. 'Q1')}

\item{cache_folder}{Path of cache folder to save all memoise files and downloaded zip files}
}
\value{
TRUE, if sucessful
}
\description{
Downloads a single file from <https://www.sec.gov/files/dera/data/financial-statement-data-sets/>.
}
\examples{
\dontrun{
flag <- download_single_zip_file(year_in = 2018,
                                 quarter_in = 'Q1',
                                 cache_folder = 'GetEdgarData-cache')
}
}
