% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_available_files}
\alias{get_available_files}
\title{Scrapes available datasets from the SEC website}
\usage{
get_available_files()
}
\value{
A dataframe with files information
}
\description{
This function find the available zip files from the SEC website <https://www.sec.gov/files/dera/data/financial-statement-data-sets/>,
returning a dataframe with the information about files.
}
\examples{

\dontrun{
df_files <- get_available_periods()
}
}
