% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edgar_fin_data.R
\name{get_edgar_fin_data}
\alias{get_edgar_fin_data}
\title{Imports data from Edgar (SEC/US)}
\usage{
get_edgar_fin_data(companies, years = seq(2009,
  lubridate::year(Sys.Date())), type_data = "yearly",
  consolidated = TRUE, do_cache = TRUE,
  cache_folder = "GetEdgarData-cache")
}
\arguments{
\item{companies}{Name of companies (see get_info_companies() for querying available names)}

\item{years}{Years of datasets (numeric vector)}

\item{type_data}{Type of dataset ('yearly' or 'quarterly')}

\item{consolidated}{Consolidade financial statements? TRUE (default) or FALSE}

\item{do_cache}{Use memoise cache system? TRUE (default) or FALSE}

\item{cache_folder}{Path of cache folder to save all memoise files and downloaded zip files}
}
\value{
A dataframe with financial statements
}
\description{
Every company trades in the US must deliver its financial statements to the SEC.
Edgar is the system that organizes such information and allows user to query the documents.
}
\details{
This function downloads data from zip files available at SEC/DERA (<https://www.sec.gov/dera/data>).
}
\examples{

\dontrun{

my.company <- 'APPLE INC'
my.years <- 2016:2018
my.type <- 'yearly'

out <- get_edgar_fin_data(companies = my.company,
                         years = my.years,
                         type.data = my.type,
                         consolidated = TRUE)
}
}
