% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_companies.R
\name{get_info_companies}
\alias{get_info_companies}
\title{Get information about available companies}
\usage{
get_info_companies(years = 2009:2018, type_data = "yearly",
  type_form = "10-K", do_cache = TRUE,
  cache_folder = "GetEdgarData-cache")
}
\arguments{
\item{years}{Years to find data}

\item{type_data}{Type of dataset ('yearly' or 'quarterly')}

\item{type_form}{The type of SEC form ('10-k' or '10-Q')}

\item{do_cache}{Use memoise cache system? TRUE (default) or FALSE}

\item{cache_folder}{Path of cache folder to save all memoise files and downloaded zip files}
}
\value{
A dataframe with information about companies
}
\description{
This functino will read the zip files and get information about available companies.
}
\examples{

\dontrun{
df_info <- get_info_companies()
}
}
