% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts-tables.R
\name{get_pre_table}
\alias{get_pre_table}
\title{Imports "pre" table}
\usage{
get_pre_table(path_to_files = tempdir())
}
\arguments{
\item{path_to_files}{Path to the folder where zip file was decompressed}
}
\value{
A dataframe
}
\description{
Imports "pre" table
}
\examples{

\dontrun{
df_raw_pre <- get_pre_table(path_to_files = tempdir())
}

}
