% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_companies.R
\name{get_sic_codes}
\alias{get_sic_codes}
\title{Get SIC codes from SIC}
\usage{
get_sic_codes(cache_folder)
}
\arguments{
\item{cache_folder}{Path of cache folder to save all memoise files and downloaded zip files}
}
\value{
A dataframe with sic codes
}
\description{
Reads and caches SIC codes from SEC (<https://www.sec.gov/info/edgar/siccodes.htm>)
}
\examples{

\dontrun{
df_sic <- get_sic_codes(cache_folder = tempdir())
}

}
