% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts-tables.R
\name{get_single_raw_table}
\alias{get_single_raw_table}
\title{Imports a single table from unzipped file}
\usage{
get_single_raw_table(path_to_files = tempdir(), type_table, cache_folder)
}
\arguments{
\item{path_to_files}{Path to the folder where zip file was decompressed}

\item{type_table}{The type of table ('num', 'sub', 'pre')}

\item{cache_folder}{Path of cache folder to save all memoise files and downloaded zip files}
}
\value{
A dataframe with the requested data
}
\description{
After unzipping the file from Edgar, this function will read a particular type of table (num, sub or pre).
}
\examples{

\dontrun{
df <- get_single_raw_table(path_to_files = tempdir(),
                           'num',
                           'GetEdgar-Cache')
}
}
