% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts-tables.R
\name{get_tag_table}
\alias{get_tag_table}
\title{Imports "tag" table}
\usage{
get_tag_table(path_to_files = tempdir())
}
\arguments{
\item{path_to_files}{Path to the folder where zip file was decompressed}
}
\value{
A dataframe
}
\description{
Imports "tag" table
}
\examples{
\dontrun{
df_raw_tag <- get_tag_table(path_to_files = tempdir())
}
}
