% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts-tables.R
\name{parse_final_table}
\alias{parse_final_table}
\title{Parses final table from zip file}
\usage{
parse_final_table(df_num, df_sub, type_data, year_in)
}
\arguments{
\item{df_num}{Dataframe with numerical data}

\item{df_sub}{Dataframe with sub table}

\item{type_data}{Type of dataset ('yearly' or 'quarterly')}

\item{year_in}{The year}
}
\value{
A dataframe with the final data
}
\description{
Parses final table from zip file
}
\examples{
\dontrun{
df_final <- parse_final_table(df_num, df_sub, 'quarterly', 2018)
}
}
