% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts-tables.R
\name{parse_num_table}
\alias{parse_num_table}
\title{Organizes a "num" table from unzipped file}
\usage{
parse_num_table(df.in)
}
\arguments{
\item{df.in}{Raw dataframe from .txt file}
}
\value{
A dataframe
}
\description{
Organizes a "num" table from unzipped file
}
\examples{

\dontrun{
df_num <- parse_num_table(raw_num_df)
}
}
