% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts-tables.R
\name{parse_sub_table}
\alias{parse_sub_table}
\title{Organizes a "sub" table from unzipped file}
\usage{
parse_sub_table(df.in)
}
\arguments{
\item{df.in}{Raw dataframe from .txt file}
}
\value{
An organized dataframe
}
\description{
Organizes a "sub" table from unzipped file
}
\examples{

\dontrun{
df_sub <- parse_sub_table(raw_sub_df)
}
}
